Only modify the chart width based on columns when showLegend is enabled
In the 2.x to 3.x upgrade we started defaulting the chart width to always be modified by columns.
Previously we only modified based on columns if showLegend was set. The result of this was a chart that was always shifted to the side due and smaller than the set view width.