Handle area charts based on a specific floor value
The purpose of this option is to provide an area chart
which start on a specific floor value and can handle reverse areas.
Instead of starting the area from the lowest value of all series,
if a serie contains values inferior to the floor value,
it will display a reverse area with a "negative" spike.