Round the individual bars up
This will consistently round the bars up. The scale function works with
fractions. Since those fractions are used to set height and y for bars, we end
up with fractions of a pixel. It is then up to the browser to round it and
browsers do that differently. If bar #1 is rounded down and bar #2 is rounded up,
we get gaps.