Generating unique ids without memory leaks
The id() function now uses a Linear Congruential Generator with a 21-bit
period so that it generates pseudo-random numbers without duplicates,
without the need of a cache holding already used values.

Given the 4-digit mantissa of this method, after around 1.6 milion ids
it will start to generate duplicates consistently, but it is still
better than a stack-overflow as the old implementation (which, by the
way came much before reaching the 1.6 milion limit).