feat: Handle the label overflow inside the Sankey chart. (#2066)
* fix: updated sankey chart component to prevent label overflow

Updated the SankeyChart component to handle overflow of the long labels from the chart. Truncated the labels based on the available space between the label's node and it's adjacent node. Added an input field 'truncateLabels' to toggle the label truncation.

* added method to handle overlapping between labels

Added a method to handle the overlapping between nodes in layer 1 and layer 2, nodes are checked with their center, if the min height difference between node's center is less than 14(label height), then they are truncated so that labels can fit inside the width available between nodes. Updated the charWidth to '6'

* handle null values for adjacentNodes